split discretization into manageable file sizes
